+++ src/redux/columnHeaderSetingsMenu/columnHeaderSettingsMenu__action.js
@@ -1,0 +1,9 @@
+export const constants = {
+    GET_SETTINGS_MENU:'GET_SETTINGS_MENU'
+}
+
+export const getSettingsMenu = ()=>{
+    return({
+        type:constants.GET_SETTINGS_MENU
+    })
+}